fix speed issues with creating csr structures.
fixes #43 -- a new constructor for SparskitBuffers class which takes the
number of nonzeros. This is typically much smaller than the heuristic
length, so it reduces the memory footprint.

fixes #51 -- two patches were applied
1) the column vector size is setto the number of nonzeros. This fixes
the large memory allocation times, and issues with out of memory errors
for large matrices (e.g. previously they were set to be size row*col
which is the size of the full dense matrix!)
2) the add functionality was revamped to first construct a coordinate
matrix and convert it to a csr matrix on finalization. This
significantly improves matrix creation times, and fixes #40.